CAR PRICE PREDICTION WITH PYTHON
CAR PRICE PREDICTION WITH PYTHON

The price of a car depends on a lot of factors like the goodwill of the brand of the car, features of the car, horsepower and the mileage it gives and many more. Car price prediction is one of the major research areas in machine learning.